added login func, acc and session management
added a login button for login and it will check if the admin is login. If not a login window will pop up, else an account management window will do so
added function to display the account management options (Edit Account, Logout, Cancel)
admin is allowed to edit their account name and password.
added and stored the admin's account information
can display the appropriate welcome message based on whether the user is an admin or a customer.
added functions to show success messages when the admin's account is updated or when they log out.